refactor(web): remove dynamic-imports barrel, import examples directly
tiptap v3 handles SSR gracefully so ssr:false is unnecessary.
Each page now imports its co-located example.tsx directly.
