+++ apps/web/src/app/editor/examples/basic-editor/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { BasicEditor as Example } from '../dynamic-imports';
+import { BasicEditor as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
