+++ apps/web/src/app/editor/examples/column-layouts/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { ColumnLayouts as Example } from '../dynamic-imports';
+import { ColumnLayouts as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
