+++ apps/web/src/app/editor/examples/custom-bubble-menu/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { CustomBubbleMenu as Example } from '../dynamic-imports';
+import { CustomBubbleMenu as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
