+++ apps/web/src/app/editor/examples/custom-extensions/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { CustomExtensions as Example } from '../dynamic-imports';
+import { CustomExtensions as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
