+++ apps/web/src/app/editor/examples/email-export/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { EmailExport as Example } from '../dynamic-imports';
+import { EmailExport as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
