+++ apps/web/src/app/editor/examples/email-theming/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { EmailThemingExample as Example } from '../dynamic-imports';
+import { EmailThemingExample as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
