+++ apps/web/src/app/editor/examples/full-email-builder/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { FullEmailBuilder as Example } from '../dynamic-imports';
+import { FullEmailBuilder as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
