+++ apps/web/src/app/editor/examples/inspector-custom/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { InspectorCustom as Example } from '../dynamic-imports';
+import { InspectorCustom as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
