+++ apps/web/src/app/editor/examples/inspector-defaults/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { InspectorDefaults as Example } from '../dynamic-imports';
+import { InspectorDefaults as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
