+++ apps/web/src/app/editor/examples/link-editing/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { LinkEditing as Example } from '../dynamic-imports';
+import { LinkEditing as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
