+++ apps/web/src/app/editor/examples/one-line-editor-full/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { OneLineEditorFull as Example } from '../dynamic-imports';
+import { OneLineEditorFull as Example } from './example';
 import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
